requirements.txt
playwright>=1.42.0
google-genai>=0.1.0
Jinja2>=3.1.3
asyncio>=3.4.3
